+++ docs/changelog.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `AutoFitLengthConverter` to the public API.
 - `lowerBound` property to `LengthUnit` interface.
 - `upperBound` property to `LengthUnit` interface.
+- `withPartsSeparator(separator)` method to `ImperialDistanceFormatter.Builder`.
 
 ### Changed
 - `Distance` and `Length` throw now `ArithmeticException` in case of failures due to overflows or math operations.
@@ -23,6 +24,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Companion objects of `SiLengthUnit` and `ImperialLengthUnit`.
 - `contains` function from `LengthUnit` interface.
 - Explicit Java support with `@Jvm*` annotations.
+- `partsSeparator` argument from `ImperialDistanceFormatter.Builder.build()` method. Use explicit `withPartsSeparator(separator)` instead.
 
 ### Fixed
 - Wrong lower meter bound of `Miles` unit.